+++ src/layers/StatsGarageLayer.cpp
@@ -1,7 +1,5 @@
 #include "StatsGarageLayer.h"
 #include <StatsDisplayAPI.h>
-#include <raydeeux.pages_api/include/PagesAPI.h>
-#include <raydeeux.pages_api/include/PageMenu.h>
 
 using namespace geode::prelude;
 
@@ -12,8 +10,6 @@ const float TOP_MARGIN = 6.f;
 float RIGHT_MARGIN = 18.f;
 // the pages api arrow scale
 float ARROW_SCALE = 0.6f;
-// the pages api arrow padding
-float ARROW_PADDING = 1.f;
 
 inline void addDefaultItem(
 	CCNode* statsMenu,
@@ -50,10 +46,12 @@ bool StatsGarageLayer::init() {
 		ColumnLayout::create()
 		->setAxisReverse(true)
 		->setCrossAxisAlignment(AxisAlignment::End)
+		->setAxisAlignment(AxisAlignment::End)
 		->setGap(15)
 	);
 	
 	this->addChild(m_fields->m_statsMenu);
+	
 	addDefaultItem(m_fields->m_statsMenu, "stars-stat", "stars", "GJ_starsIcon_001.png", "6", 0.54f);
 	addDefaultItem(m_fields->m_statsMenu, "moons-stat", "moons", "GJ_moonsIcon_001.png", "28", 0.54f);
 	addDefaultItem(m_fields->m_statsMenu, "gold-coins-stat", "coins", "GJ_coinsIcon_001.png", "8", 0.51f);
@@ -61,7 +59,6 @@ bool StatsGarageLayer::init() {
 	addDefaultItem(m_fields->m_statsMenu, "orbs-stat", "orbs", "currencyOrbIcon_001.png", "14", 0.54f);
 	addDefaultItem(m_fields->m_statsMenu, "diamonds-stat", "diamonds", "GJ_diamondsIcon_001.png", "13", 0.6f);
 	addDefaultItem(m_fields->m_statsMenu, "diamond-shards-stat", "diamond-shards", "currencyDiamondIcon_001.png", "29", 0.54f);
-
 	
 	float bottomMargin = CCSprite::createWithSpriteFrameName("GJ_sideArt_001.png")->getContentHeight();
 	float arrowSize = ARROW_SCALE * CCSprite::createWithSpriteFrameName("GJ_arrow_02_001.png")->getContentWidth();
@@ -69,83 +66,104 @@ bool StatsGarageLayer::init() {
 	// the maximum height that our elements should cover, including the arrows and their paddings as "elements"
 	float maxHeight = winSize.height - TOP_MARGIN - bottomMargin;
 	// the actual maximum of the content, excluding the arrows
-	m_fields->m_statsMenu->setContentSize({0, maxHeight - 2*ARROW_PADDING - 2*arrowSize});
+	m_fields->m_statsMenu->setContentSize({0, maxHeight});
 	m_fields->m_statsMenu->setAnchorPoint(CCPoint{0.5f, 1.f});
-	m_fields->m_statsMenu->setPositionX(winSize.width - RIGHT_MARGIN);
+	m_fields->m_statsMenu->setPosition(CCPoint{winSize.width - RIGHT_MARGIN, winSize.height - TOP_MARGIN});
+	
+	// set up the arrow buttons for pages
+	CCSprite* prevSprite = CCSprite::createWithSpriteFrameName("GJ_arrow_02_001.png");
+	CCSprite* nextSprite = CCSprite::createWithSpriteFrameName("GJ_arrow_02_001.png");
+	nextSprite->setFlipX(true);
+	prevSprite->setRotation(90);
+	nextSprite->setRotation(90);
+	prevSprite->setScale(ARROW_SCALE);
+	nextSprite->setScale(ARROW_SCALE);
+	m_fields->m_prevArrow = CCMenuItemSpriteExtra::create(prevSprite, this, menu_selector(StatsGarageLayer::switchPage));
+	m_fields->m_nextArrow = CCMenuItemSpriteExtra::create(nextSprite, this, menu_selector(StatsGarageLayer::switchPage));
+	m_fields->m_nextArrow->setID("next-arrow"_spr);
+	m_fields->m_prevArrow->setID("prev-arrow"_spr);
+	
+	// set the Z order so that the arrows are always at the start and end
+	m_fields->m_statsMenu->addChild(m_fields->m_prevArrow, INT_MIN, -1);
+	m_fields->m_statsMenu->addChild(m_fields->m_nextArrow, INT_MAX, 1);
 
-	if (PagesAPI::isLoaded()) {
-    PageMenu* pageMenu = static_cast<PageMenu*>(m_fields->m_statsMenu); 
-		pageMenu->setPaged(ELEMENTS_PER_PAGE, PageOrientation::VERTICAL, maxHeight, ARROW_PADDING);
-		pageMenu->setButtonScale(ARROW_SCALE);
-		schedule(schedule_selector(StatsGarageLayer::moveMenuForArrows));
-	}
+	// this checks for newly added children and updates paging if they exist
+	schedule(schedule_selector(StatsGarageLayer::pageChildren));
 
 	return true;
 }
 
-CCNode* getFirstVisibleChild(CCNode* parent) {
-	for (auto child : parent->getChildrenExt())
-		if (child->isVisible())
-			return child;
-	return nullptr;
-}
+// the below 2 implement custom paging
 
-CCNode* getLastVisibleChild(CCNode* parent) {
-	CCArrayExt<CCNode*> arrayExt = parent->getChildrenExt();
-	if (arrayExt.size() == 0)
-		return nullptr;
-	for (size_t i = arrayExt.size()-1; i >= 0; i--)
-		if (arrayExt[i]->isVisible())
-			return arrayExt[i];
-	return nullptr;
-}
-
-// this is a hack to move the node properly so it works with the way pages api works
-// another issue this has it that if you switch a page and it needs to move the menu
-// the arrows will visually jitter for a frame
-// i do not know how to fix this without modifying pages api to add button press callbacks
-// so we can do this instantly on button press
-// it would genuinely be much easier to implement your own paging at this point
-void StatsGarageLayer::moveMenuForArrows(float) {
-  auto winSize = CCDirector::get()->getWinSize();
-	CCNode* firstChild = getFirstVisibleChild(m_fields->m_statsMenu);
-	CCNode* lastChild = getLastVisibleChild(m_fields->m_statsMenu);
-	if (!firstChild || !lastChild) return;
-	CCNode* firstLabel = firstChild->getChildByIndex(1);
-	CCNode* lastLabel = lastChild->getChildByIndex(1);
-	if (!firstLabel || !lastLabel) return;
-
-	float topLabelMaxY = firstChild->convertToWorldSpace(CCPoint{0, firstLabel->boundingBox().getMaxY()}).y;
-	float bottomLabelMinY = lastChild->convertToWorldSpace(CCPoint{0, lastLabel->boundingBox().getMinY()}).y;
-	
-	float newTopLabelMaxY;
-	if (CCNode* pageMenu = m_fields->m_statsMenu->getParent()->getChildByID("stats-menu-navigation-menu"_spr)) {
-		CCNode* previous = pageMenu->getChildByID("prev-button");
-		CCNode* next = pageMenu->getChildByID("next-button");
-		if (!previous || !next) return;
-
-		float previousMaxY = pageMenu->convertToWorldSpace(CCPoint{0, previous->boundingBox().getMaxY()}).y;
-		float nextMinY = pageMenu->convertToWorldSpace(CCPoint{0, next->boundingBox().getMinY()}).y;
-		
-		float prevDelta = previousMaxY - topLabelMaxY - ARROW_PADDING - previous->getScaledContentHeight();
-		float nextDelta = bottomLabelMinY - nextMinY - ARROW_PADDING - next->getScaledContentHeight();
-		previous->setPositionY(previous->getPositionY() - prevDelta);
-		next->setPositionY(next->getPositionY() + nextDelta);
-		// if the current top label's max y is topLabelMaxY
-		// we want to move everything so that
-		// winSize.height - newTopLabelMaxY = previous->getContentHeight() + ARROW_PADDING + TOP_MARGIN
-		// newTopLabelMaxY = winSize.height - previous->getContentHeight() - ARROW_PADDING - TOP_MARGIN
-		// we don't multiply the height by button scale because the scale is on the inner sprite,
-		// not the button itself, so its content height is adjusted to the scale already
-		// if we don't have arrows, we exclude them and their padding from the equation and thus get the 2nd branch
-		// newTopLabelMaxY = winSize.height - TOP_MARGIN
-		newTopLabelMaxY = winSize.height - previous->getContentHeight() - ARROW_PADDING - TOP_MARGIN;
-	} else {
-		// if the arrow menu does not exist, then we only have a single page and do not need to
-		// keep track of anything else, just move the menu to align with the top margin
-		newTopLabelMaxY = winSize.height - TOP_MARGIN;
+void StatsGarageLayer::switchPage(CCObject* sender) {
+	if (!m_fields->m_statsMenu) return;
+	CCNode* button = static_cast<CCNode*>(sender);
+	if (m_fields->m_statsMenu->getChildrenCount() < 2) {
+		m_fields->m_currentPage = 0;
+		return;
 	}
-	// the amount we need to move the menu so that the top label max y becomes newTopLabelMaxY
-	float topLabelMaxYDelta = topLabelMaxY - newTopLabelMaxY;
-	m_fields->m_statsMenu->setPositionY(m_fields->m_statsMenu->getPositionY() - topLabelMaxYDelta);
+	if (m_fields->m_statsMenu->getChildrenCount() == 2) m_fields->m_requestedPage = 0;
+
+	int actualChildren = m_fields->m_statsMenu->getChildrenCount() - 2;
+	int pageDelta = button->getTag();
+	int maxPage = (actualChildren - 1)/ELEMENTS_PER_PAGE;
+	if (m_fields->m_currentPage + pageDelta < 0) m_fields->m_requestedPage = maxPage;
+	else if (m_fields->m_currentPage + pageDelta > maxPage) m_fields->m_requestedPage = 0;
+	else m_fields->m_requestedPage = m_fields->m_currentPage + pageDelta;
+}
+
+void StatsGarageLayer::pageChildren(float) {
+	if (!m_fields->m_statsMenu) return;
+
+	if (m_fields->m_statsMenu->getChildrenCount() < 2) return;
+	int actualChildren = m_fields->m_statsMenu->getChildrenCount() - 2;
+	// don't unnecessarily update, only do it if anything actually changed
+	if (
+		m_fields->m_requestedPage == m_fields->m_currentPage &&
+		actualChildren == m_fields->m_previousActualChildren.size()
+	) {
+		bool allSame = true;
+		for (size_t i = 1; i < m_fields->m_statsMenu->getChildrenCount()-1; i++)
+			if (m_fields->m_statsMenu->getChildByIndex(i) != m_fields->m_previousActualChildren[i-1])
+				allSame = false;
+		if (allSame) return;
+	}
+
+	m_fields->m_previousActualChildren.clear();
+	m_fields->m_currentPage = m_fields->m_requestedPage;
+
+	m_fields->m_prevArrow->setVisible(m_fields->m_statsMenu->getChildrenCount() > ELEMENTS_PER_PAGE + 2);
+	m_fields->m_nextArrow->setVisible(m_fields->m_statsMenu->getChildrenCount() > ELEMENTS_PER_PAGE + 2);
+	
+	// set up the current page and save the previous children
+	int maxPage = actualChildren == 0 ? 0 : (actualChildren - 1)/ELEMENTS_PER_PAGE;
+	if (m_fields->m_currentPage > maxPage) m_fields->m_currentPage = maxPage; 
+	for (size_t i = 1; i < m_fields->m_statsMenu->getChildrenCount()-1; i++) {
+		CCNode* child = m_fields->m_statsMenu->getChildByIndex(i);
+		m_fields->m_previousActualChildren.push_back(child);
+		child->setVisible((i - 1)/ELEMENTS_PER_PAGE == m_fields->m_currentPage);
+	}
+	
+	m_fields->m_statsMenu->updateLayout();
+	
+	// adjust position of menu. ideally we would not have to do this, but oh well
+	auto winSize = CCDirector::get()->getWinSize();
+	if (maxPage == 0 && actualChildren != 0) {
+		// the containers have 0 height, and that's in the api now
+		// so i have to do this to not touch the api
+		// on the next major version, remove this and give the containers actual height
+		// depending on the children; may have to adjust the gap as a result
+		// i will use the label's top position to set the menu's position
+		CCNode* firstChild = m_fields->m_statsMenu->getChildByIndex(1);
+		CCNode* firstLabel = firstChild->getChildByIndex(1);
+		if (!firstLabel) return;
+
+		float topLabelMaxY = firstChild->convertToWorldSpace(CCPoint{0, firstLabel->boundingBox().getMaxY()}).y;
+		float newTopLabelMaxY = winSize.height - TOP_MARGIN;
+		float topLabelMaxYDelta = topLabelMaxY - newTopLabelMaxY;
+		m_fields->m_statsMenu->setPositionY(m_fields->m_statsMenu->getPositionY() - topLabelMaxYDelta);
+	} else {
+		// otherwise we can do the sensible thing of keeping the position as it's supposed to be
+		m_fields->m_statsMenu->setPositionY(winSize.height - TOP_MARGIN);
+	}
 }